+++ Reach/test/utopia/reach/test/SelectionListTest.scala
@@ -25,7 +25,7 @@ object SelectionListTest extends App
 	import ReachTestContext._
 	
 	// Data
-	val contentPointer = new PointerWithEvents(Vector(1, 2, 3, 4, 5))
+	val contentPointer = new PointerWithEvents(Vector(1, 2, 3, 4, 5, 6, 7))
 	val valuePointer = new PointerWithEvents[Option[Int]](Some(2))
 	
 	val mainBg = colors.gray.default